Adds endpoint pitches/stats

+++ app/models/user_token.rb
@@ -3,7 +3,11 @@ class UserToken < ActiveRecord::Base
   belongs_to :user
 
   def delete_expired_tokens
+    return if ! self.user_id.present?
+
     user = User.find( self.user_id )
+    return if ! user.present?
+    
     if user.user_tokens.count > 5
       user.user_tokens.first.destroy
     end
